feat: reputation & opcode validation & other fixes

+++ src/bundler/rpc/services/BundlingService.ts
@@ -1,7 +1,6 @@
 import { NetworkNames } from 'etherspot';
-import { providers } from 'ethers';
+import { BigNumber, ethers, providers } from 'ethers';
 import logger from 'app/logger';
-import { EntryPointContract } from 'app/bundler/contracts';
 import { MempoolService } from './MempoolService';
 import {
   UserOpValidationResult,
@@ -9,6 +8,10 @@ import {
 } from './UserOpValidation';
 import { MempoolEntry } from '../entities/MempoolEntry';
 import config from 'app/config';
+import { EntryPoint__factory } from 'app/bundler/contracts/factories';
+import { getAddr } from 'app/bundler/utils';
+import { ReputationService } from './ReputationService';
+import { ReputationStatus } from 'app/@types';
 
 export class BundlingService {;
   constructor(
@@ -16,6 +19,7 @@ export class BundlingService {;
     private provider: providers.JsonRpcProvider,
     private mempoolService: MempoolService,
     private userOpValidationService: UserOpValidationService,
+    private reputationService: ReputationService
   ) {}
 
   async sendBundle(bundle: MempoolEntry[]): Promise<void> {
@@ -23,16 +27,16 @@ export class BundlingService {;
       return;
     }
     const entryPoint = bundle[0]!.entryPoint;
-    const entryPointContract = new EntryPointContract(
+    const entryPointContract = EntryPoint__factory.connect(
+      entryPoint,
       this.provider,
-      entryPoint
     );
     const wallet = config.getEntryPointRelayer(this.network, entryPoint)!;
     const beneficiary = config.getEntryBeneficiary(this.network, entryPoint)!;
     try {
-      let txRequest = entryPointContract.encodeHandleOps(
-        bundle.map(entry => entry.userOp),
-        beneficiary
+      let txRequest = entryPointContract.interface.encodeFunctionData(
+        'handleOps',
+        [ bundle.map(entry => entry.userOp), beneficiary ]
       );
       const tx = await wallet.sendTransaction({
         to: entryPoint,
@@ -50,22 +54,79 @@ export class BundlingService {;
         logger.warn(`Failed handleOps, but non-FailedOp error ${err}`);
         return;
       }
-      const { index } = err.errorArgs;
+      const {
+        index,
+        paymaster,
+        reason
+      } = err.errorArgs;
       const entry = bundle[index]!;
-      await this.mempoolService.remove(entry);
+      if (paymaster !== ethers.constants.AddressZero) {
+        this.reputationService.crashedHandleOps(paymaster);
+      } else if (typeof reason === 'string' && reason.startsWith('AA1')) {
+        const factory = getAddr(entry.userOp.initCode);
+        if (factory) {
+          this.reputationService.crashedHandleOps(factory);
+        }
+      } else {
+        await this.mempoolService.remove(entry);
+        logger.warn(`Failed handleOps sender=${entry.userOp.sender}`);
+      }
     }
   };
 
   async createBundle(): Promise<MempoolEntry[]> {
     // TODO: support multiple entry points
     //       filter bundles by entry points
-    // TODO: add total gas cap
-    // TODO: reputation
     const entries = await this.mempoolService.getSortedOps();
     const bundle: MempoolEntry[] = [];
 
+    const paymasterDeposit: { [key: string]: BigNumber } = {};
+    const stakedEntityCount: { [key: string]: number } = {};
     const senders = new Set<string>();
     for (const entry of entries) {
+      const paymaster = getAddr(entry.userOp.paymasterAndData);
+      const factory = getAddr(entry.userOp.initCode);
+
+      if (paymaster) {
+        const paymasterStatus = await this.reputationService.getStatus(paymaster);
+        if (paymasterStatus === ReputationStatus.BANNED) {
+          await this.mempoolService.remove(entry);
+          continue;
+        } else if (
+          paymasterStatus === ReputationStatus.THROTTLED ||
+          (stakedEntityCount[paymaster] || 0) > 1
+        ) {
+          logger.debug('skipping throttled paymaster', {
+            metadata: {
+              sender: entry.userOp.sender,
+              nonce: entry.userOp.nonce,
+              paymaster
+            }
+          });
+          continue;
+        }
+      }
+
+      if (factory) {
+        const deployerStatus = await this.reputationService.getStatus(factory);
+        if (deployerStatus === ReputationStatus.BANNED) {
+          await this.mempoolService.remove(entry);
+          continue;
+        } else if (
+          deployerStatus === ReputationStatus.THROTTLED ||
+          (stakedEntityCount[factory] || 0) > 1
+        ) {
+          logger.debug('skipping throttled factory', {
+            metadata: {
+              sender: entry.userOp.sender,
+              nonce: entry.userOp.nonce,
+              factory
+            }
+          });
+          continue;
+        }
+      }
+
       if (senders.has(entry.userOp.sender)) {
         logger.debug('skipping already included sender', {
           metadata: {
@@ -87,6 +148,31 @@ export class BundlingService {;
         await this.mempoolService.remove(entry);
         continue;
       }
+      
+      // TODO: add total gas cap
+      const entryPointContract = EntryPoint__factory.connect(
+        entry.entryPoint,
+        this.provider,
+      );
+      if (paymaster) {
+        if (!paymasterDeposit[paymaster]) {
+          paymasterDeposit[paymaster] = await entryPointContract.balanceOf(paymaster);
+        }
+        if (paymasterDeposit[paymaster]?.lt(validationResult.returnInfo.prefund)) {
+          // not enough balance in paymaster to pay for all UserOps
+          // (but it passed validation, so it can sponsor them separately
+          continue;
+        }
+        stakedEntityCount[paymaster] = (stakedEntityCount[paymaster] || 0) + 1;
+        paymasterDeposit[paymaster] = BigNumber.from(
+          paymasterDeposit[paymaster]?.sub(validationResult.returnInfo.prefund)
+        );
+      }
+
+      if (factory) {
+        stakedEntityCount[factory] = (stakedEntityCount[factory] || 0) + 1;
+      }
+
       senders.add(entry.userOp.sender);
       bundle.push(entry);
     }
